refactor: createHashObject 함수 GeetUtil.kt로 옮김
